requirements.txt
numpy
matplotlib
json
pandas
pyyaml
IPython
nbformat
nbconvert==6.1.0